Fix compiler warnings

+++ src/config.cpp
@@ -43,13 +43,13 @@ e_config_load_result load_config_file()
 
 	if (config_file!=NULL)
 	{
-		const int k_buffer_size= 1024;
+		const size_t k_buffer_size = 1024;
 		char buffer[k_buffer_size];
-		int bytes_read= 0;
-		int buffer_space= k_buffer_size-1;
-		int bytes_to_read;
-		char *current_position= buffer;
-		int left_over_length= 0;
+		size_t bytes_read = 0;
+		size_t buffer_space = k_buffer_size - 1;
+		size_t bytes_to_read;
+		char *current_position = buffer;
+		size_t left_over_length = 0;
 
 
 		do
